raptai-fastapi/pyproject.toml
[project]
name = "raptai-pro"
version = "0.1.0"
description = ""
authors = [
    {name = "balajihambeere",email = "balaji.hambere@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "databases[asyncpg] (>=0.9.0,<0.10.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [{ include = "app", from = "src" }]
